% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_wrappers.R
\name{calf_exact_binary_subset}
\alias{calf_exact_binary_subset}
\title{calf_exact_binary_subset}
\usage{
calf_exact_binary_subset(data, nMarkers, nCase, nControl, times = 1,
  margin = NULL, optimize = "pval", verbose = FALSE)
}
\arguments{
\item{data}{Matrix or data frame. First column must contain case/control dummy coded variable (if targetVector = "binary"). Otherwise, first column must contain real number vector corresponding to selection variable (if targetVector = "nonbinary"). All other columns contain relevant markers.}

\item{nMarkers}{Maximum number of markers to include in creation of sum.}

\item{nCase}{Numeric. A value indicating the number of case data to use.}

\item{nControl}{Numeric. A value indicating the number of control data to use.}

\item{times}{Numeric. Indicates the number of replications to run with randomization.}

\item{margin}{Real number from 0 to 1. Indicates the amount a potential marker must improve the target criterion (Pearson correlation or p-value) in order to add the marker.}

\item{optimize}{Criteria to optimize if targetVector = "binary." Indicate "pval" to optimize the p-value corresponding to the t-test distinguishing case and control. Indicate "auc" to optimize the AUC.}

\item{verbose}{Logical. Indicate TRUE to print activity at each iteration to console. Defaults to FALSE.}
}
\value{
A data frame containing the chosen markers and their assigned weight (-1 or 1)

The AUC value for the classification. If multiple replications are requested, this will be a data.frame containing all AUCs across replications.

aucHist A histogram of the AUCs across replications.
}
\description{
Runs Coarse Approximation Linear Function on a random subset of binary data provided, with the ability to precisely control the number of case and control data used.
}
\examples{
calf_exact_binary_subset(data = CaseControl, nMarkers = 6, nCase = 5, nControl = 8, times = 5)
}
