% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALF-data.R
\docType{data}
\name{CalfBin}
\alias{CalfBin}
\title{Example binary data for CALF algorithm}
\format{A data.table with the first column giving the response variable, 
and the remaining columns giving predictor variables.}
\usage{
data(CalfBin)
}
\description{
This data contains standardized data for 135 blood markers for 72 individuals 
who are distinguished as control (\code{CalfBin$NCvC == 0}) & case 
(\code{CalfBin$NCvC == 0}).

The response variable (designating case/control) is given by the first 
column, \code{NCvC}.The remaining columns, (i.e. \code{M---}) are different
blood markers (the predictor variables).

This data comes from the 
[North American Prodrome Longitudinal Study](https://napls.ucsf.edu).
}
\keyword{datasets}
