% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALF-data.R
\docType{data}
\name{CalfCnt}
\alias{CalfCnt}
\title{Example continuous data for CALF algorithm}
\format{A data.table with the first column giving the response variable, 
and the remaining columns giving predictor variables.}
\usage{
data(CalfCnt)
}
\description{
This data contains standardized expression for 132 miRNAs in blood for 24 
people and their percentage of blood monocytes.

The response variable (fraction of monocytes) is given by the first 
column, \code{mono}.The remaining columns, (i.e. \code{M---}) are different
blood miRNAs (the predictor variables).

This data comes from the 
\href{https://napls.ucsf.edu}{North American Prodrome Longitudinal Study}.
}
\keyword{datasets}
