% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALF-data.R
\docType{data}
\name{CaseControl}
\alias{CaseControl}
\title{Example data containing case and control data}
\format{A data frame with 136 marker variables and 68 individuals.}
\usage{
CaseControl
}
\description{
This data contains 136 marker variables for 68 individuals who are distinguished as case/control.
}
\keyword{datasets}

