% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_wrappers.R
\name{write.calf}
\alias{write.calf}
\title{write.calf}
\usage{
write.calf(x, filename)
}
\arguments{
\item{x}{A CALF data frame.}

\item{filename}{The output filename}
}
\description{
Writes output of the CALF dataframe
}
