% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_wrappers.R
\name{write.calf_randomize}
\alias{write.calf_randomize}
\title{write.calf_randomize}
\usage{
write.calf_randomize(x, filename)
}
\arguments{
\item{x}{A CALF randomize data frame.}

\item{filename}{The output filename}
}
\description{
Writes output of the CALF randomize dataframe
}
