% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calf_wrappers.R
\name{write.calf_subset}
\alias{write.calf_subset}
\title{write.calf_subset}
\usage{
write.calf_subset(x, filename)
}
\arguments{
\item{x}{A CALF subset data frame.}

\item{filename}{The output filename}
}
\description{
Writes output of the CALF subset dataframe
}
