% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALF.R
\name{calf}
\alias{calf}
\title{calf}
\usage{
calf(data, nMarkers)
}
\arguments{
\item{data}{Matrix or data frame. First column must contain case/control dummy coded variable; all other columns contain relevant markers}

\item{nMarkers}{Maximum number of markers to include in creation of sum}
}
\value{
A data frame containing the chosen markers and their assigned weight (-1 or 1)
}
\description{
Coarse approximation linear function
}
\examples{
calf(data = CaseControl, nMarkers = 6)
}

