% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calfWrappers.R
\name{calf}
\alias{calf}
\title{calf}
\usage{
calf(
  resp,
  pred,
  maxPred = ncol(pred),
  score = "pcc",
  margin = NULL,
  verbose = FALSE
)
}
\arguments{
\item{resp}{Response vector; binary or continuous.}

\item{pred}{Matrix or data.frame with predictor variables.}

\item{maxPred}{Maximum number of predictors to include in the model.}

\item{score}{Either "auc", "pcc", or "wts"; see details for more information.}

\item{margin}{Real number from 0 to 1, the required improvement in score 
include an additional predictor; see details.}

\item{verbose}{Logical. When TRUE activity is printed for each iteration.}
}
\value{
A calf object with the following list items:
\describe{
  \item{\code{predInd}}{Index of selected predictors in \code{pred} in the 
  order of selection}
  \item{\code{predName}}{Name of the selected predictors in \code{pred} in
  the order of selection}
  \item{\code{weight}}{The weight value for the selected predictors in 
  the order of selection}
  \item{\code{trainVec}}{The best score value at each iteration with the
  addition of the new predictor given in predInd; included to monitor
  the training progress}
  \item{\code{score}}{The score function used to calculate the scores in
  \code{trainVec}}
  \item{\code{maxPred}}{The maximum number of predictors allowed to be 
  included in the model}
  \item{\code{margin}}{The required improvement in score to continue 
  adding parameters to the model}
}
}
\description{
Coarse Approximation Linear Function
}
\details{
\code{resp} must be logical for binary score functions, i.e. 
\code{score = "auc"} or \code{score = "wts"}. Otherwise, \code{resp} is 
considered continuous. 

Need to add lots here...
}
\examples{

data(CalfBin)
calf(resp = as.logical(CalfBin[[1]]), pred = CalfBin[ , -1], maxPred = 5)

}
