% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreFuncs.R
\name{scoreFuncs}
\alias{scoreFuncs}
\alias{calfScoreAUC}
\alias{calfScoreWTS}
\alias{calfScorePCC}
\title{CALF score functions}
\usage{
calfScoreAUC(resp, pred)

calfScoreWTS(resp, pred)

calfScorePCC(resp, pred)
}
\arguments{
\item{resp}{Response vector}

\item{pred}{Predictor vector}
}
\description{
Available CALF score functions
}
\details{
\code{resp} MUST be logical for \code{calfScoreAUC} and \code{calfScoreWTS}.

Can add more later...
}
\section{Functions}{
\itemize{
\item \code{calfScoreAUC}: Calculate score as the area under the receivor 
operating curve; only for binary response

\item \code{calfScoreWTS}: Calculate score as the p-value from Welch's t-test; 
only for binary response

\item \code{calfScorePCC}: Calculate score as the absolute value of Pearson's 
correlation coefficient
}}

